rename more files from policy to configuration
